stream cleanups

+++ src/engine/macutils.mm
@@ -1,7 +1,7 @@
 #import <Cocoa/Cocoa.h>
 
-char *mac_pasteconsole(int *cblen)
-{
+char *mac_pasteconsole(size_t *cblen)
+{	
     NSPasteboard *pasteboard = [NSPasteboard generalPasteboard];
     NSString *type = [pasteboard availableTypeFromArray:[NSArray arrayWithObject:NSStringPboardType]];
     if(type != nil)
